kernel: Boost whenever a zygote-forked process becomes a top app
Boost to the max for 1250 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

[diepquynh]: Adapted to 4.14

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>